pyproject.toml
[project]
name = "personaplex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "huggingface-hub>=1.6.0",
    "python-dotenv>=1.2.2",
    "sentencepiece>=0.2.1",
]
